Updated project URLs in the About window.  Fixed compiler warning.  Renamed TSamplePlane to TDataBuffer, TSampleBuffer to TPlanarBuffer.  Added initial stubs for subtitle support.
git-svn-id: https://svn.code.sf.net/p/apprenticevideo/code/trunk@324 28900df4-563f-46a4-9db8-70c0268b8f86

+++ apprenticevideo/yaePlaylistWidget.cpp
@@ -352,8 +352,8 @@ namespace yae
   PlaylistWidget::sizeHint() const
   {
     int w = minimumWidth();
-    int h = (kGroupNameHeight * numShownGroups_ +
-             kGroupItemHeight * numShown_);
+    int h = int(kGroupNameHeight * numShownGroups_ +
+                kGroupItemHeight * numShown_);
 
     return QSize(w, h);
   }
